Dashboard analatics monthly transactions and enrollements by courses

+++ src/main/java/com/dangeacademy/dto/DashboardSummaryDTO.java
@@ -8,9 +8,9 @@ public class DashboardSummaryDTO {
     private Double totalRevenue;
     private Long totalStudents;
     private Long totalCourses;
-    private int totalTransactions;
+    private Long totalTransactions;
 
-    public DashboardSummaryDTO(Double totalRevenue, Long totalStudents, Long totalCourses , int totalTransactions) {
+    public DashboardSummaryDTO(Double totalRevenue, Long totalStudents, Long totalCourses , Long totalTransactions) {
         this.totalRevenue = totalRevenue;
         this.totalStudents = totalStudents;
         this.totalCourses = totalCourses;
